Fix input lag.
Every frame, the code first moved the snake, then registered input, then
updated the screen.

And for some reason, updating the screen waits until it's time to draw
the next frame.

An unfortunate consequence of this is that the game feels laggy – you
have to wait until next frame to get your input processed, which takes
100ms (because the timer is set to 10fps) – more than enough
for the player to notice.

There's a simple fix, though – process input at the start of every
frame, so that the snake responds immediately.

Also, a small suggestion: run the clock in the main loop, and not in
some function – it'll be easier to see that the whole game runs at
10fps, not just the screen update. Or extract the waiting to a separate
function – regardless, it's not really related to updating the screen.