Make a proper python module out of this app.
Up to now, this game was just a bunch of files. This is fine for very
small apps, but quickly becomes problematic, if not for other reasons,
because it makes the app uninstallable.

A python module is basically a directory with an __init__.py file.
It allows you to use relative imports, i.e.

    from .settings import Settings

Thanks to this, if there's ever a module named "settings" outside of
your application, your file won't shadow it. If you want to see what I
mean by yourself, create a `pygame.py` before applying these changes and
see what happens. Proper structure is important, and using modules is a
part of it.

Now, in order for python to recognize it as a module, you have to
execute it as such. There are multiple ways of doing that, but the
simplest one is:

    python3 -m snake_clone

This executes the __main__.py file, so I've had to move the startup code
to it.

Oh, and it has to be in a subdirectory because of the way python's
setuptools work. But more on that later.