Make the game installable.
Now, for another method of running a module – installing it.

    pip install --user -e .

Running the above will install the game in a way that it will be always
up to date with your code.

The `setup.py` file is something that one usually copies from the
internet or another project, an changes the intereting bits.

I have included a really minimal example here, there's a *lot* of fields
you'd want to fill if you were publishing it.

There's a convention for specyfing libraries your project depends on in
python: the `requirements.txt` file. It can be used to just install the
dependencies using:

    pip install -r requirements.txt

But the `setup.py` I've included also reads it. It makes others' lives
easier if you specify your dependencies like that. And yours, if you
don't want to remember and manually install whatever you depend on.

Finally, on the matter of entry points – it's a somewhat complex, and
rarely useful concept in general, but there's one often used aspect:
console_scripts. This is what allows you to create an executable to run
for your package.

The way it works is rather simple, you tell it the name of the script,
and the function to call. In this case the name is `snake-clone`, and
the function is `run` in the `snake_clone.__main__` module (I forgot to
mention – files are also python modules).

Once installed, you should be able to just run `snake-clone` from your
commandline. If not, runnign this, then restarting the console will probably help:

    echo 'export PATH=$HOME/.local/bin:$PATH` >> .bashrc

This is not related to python at all, it just configures your system to
recognize packages installed for your user (instead of globally).